requirements.txt
absl-py==0.8.1
astor==0.8.0
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
gast==0.3.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.10.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
opencv-contrib-python
gunicorn
object-detection
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.17.2
object-detection
Pillow==6.2.0
protobuf==3.10.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
